QT Build - P2P Flag, IWebSocket, IHTTPClient Abstraction, Relay Privacy

+++ desktop/QuicConnection.cpp
@@ -164,7 +164,7 @@ void QuicConnection::onIceStateChanged(int state) {
 
         // Find a free UDP port for QUIC
         QUdpSocket probe;
-        probe.bind(QHostAddress::AnyIPv4, 0);
+        probe.bind(QHostAddress::AnyIPv4, quint16(0));
         m_localQuicPort = probe.localPort();
         probe.close();
